+++ Sources/SuperwallKit/Logger/Logger.swift
@@ -6,54 +6,6 @@
 //
 
 import Foundation
-import CloudKit
-
-/// Specifies the detail of the logs returned from the SDK to the console.
-@objc(SWKLogLevel)
-public enum LogLevel: Int, CustomStringConvertible {
-  /// Prints all logs from the SDK to the console. Useful for debugging your app if something isn't working as expected.
-  case debug = 10
-  /// Prints errors, warnings, and useful information from the SDK to the console.
-	case info = 20
-  /// Prints errors and warnings from the SDK to the console.
-	case warn = 30
-  /// Only prints errors from the SDK to the console.
-	case error = 40
-
-  /// The string value of the log level
-	public var description: String {
-		switch self {
-		case .debug: return "DEBUG"
-		case .info: return "INFO"
-		case .warn: return "WARN"
-		case .error: return "ERROR"
-		}
-	}
-}
-
-/// The possible scope of logs to print to the console.
-public enum LogScope: String {
-	case localizationManager
-	case bounceButton
-  case coreData
-  case configManager
-	case debugManager
-	case debugViewController
-	case localizationViewController
-	case gameControllerManager
-	case device
-	case network
-	case paywallEvents
-	case productsManager
-	case storeKitManager
-	case events
-	case superwallCore
-	case paywallPresentation
-	case paywallTransactions
-	case paywallViewController
-	case cache
-	case all
-}
 
 enum Logger {
 	static func shouldPrint(